Total of working hours is pass into TimeSheet
