+++ TimeKeeping/schedValidator/checkSched.py
@@ -6,7 +6,7 @@ from mysql.connector import Error
 import logging
 
 from PyQt5.QtCore import QDate, Qt, QTime
-from PyQt5.QtWidgets import QDialog, QApplication, QFileDialog, QTableWidgetItem, QDateEdit, QLabel, QPushButton, QTableWidget, QMainWindow, QLineEdit
+from PyQt5.QtWidgets import QDialog, QApplication, QFileDialog, QTableWidgetItem, QDateEdit, QLabel, QPushButton, QTableWidget, QMainWindow, QLineEdit, QMessageBox, QHeaderView
 from PyQt5.uic import loadUi
 
 # Configure the logger
@@ -23,7 +23,6 @@ def resource_path(relative_path):
 def create_connection():
     try:
         connection = mysql.connector.connect(
-            #host='127.0.0.1',
             host='localhost',
             database='timekeeping',
             user='root',
@@ -43,7 +42,7 @@ class chkSched(QDialog):
     def __init__(self, data):
         super(chkSched, self).__init__()
         self.setFixedSize(731, 405)
-        ui_file = (resource_path("TimeKeeping\\schedValidator\\Schedule.ui"))
+        ui_file = resource_path("TimeKeeping\\schedValidator\\Schedule.ui")
         loadUi(ui_file, self)
 
         self.data = data
@@ -51,39 +50,16 @@ class chkSched(QDialog):
         self.populate_schedule_with_data(data)
 
     def populate_schedule_with_data(self, data):
-        (empNum, bioNum, empName, trans_date, checkIn, checkOut, sched) = data
+        (empNum, bioNum, empName, trans_date, checkIn, checkOut, sched, total_hours) = data
 
         self.empNameTxt.setText(empName)
         self.bioNumTxt.setText(bioNum)
         self.dateOfWork.setDate(QDate.fromString(trans_date, "yyyy-MM-dd"))
         self.timeInTxt.setText(checkIn)
         self.timeOutTxt.setText(checkOut)
+        self.hoursWorkedTxt.setText(str(total_hours))
         self.typeDayTxt.setText(self.getTypeOfDate(trans_date))
-        self.hoursWorkedTxt.setText(str(self.getTotalHoursWorked(checkIn, checkOut)))
         self.typeOfDayCb.setCurrentText(self.getTypeOfDate(trans_date))
-
-    def getTotalHoursWorked(self, time_start, time_end):
-        if time_start == 'Missing' or time_end == 'Missing':
-            return "Unknown"
-
-        timeIn = QTime.fromString(time_start, "HH:mm:ss")
-        timeOut = QTime.fromString(time_end, "HH:mm:ss")
-
-        # Converting time into seconds
-        seconds_in_a_day = 24 * 60 * 60
-        time_in_seconds = (timeIn.hour() * 3600) + (timeIn.minute() * 60) + timeIn.second()
-        time_out_seconds = (timeOut.hour() * 3600) + (timeOut.minute() * 60) + timeOut.second()
-
-        # Handle crossing midnight
-        if time_out_seconds < time_in_seconds:
-            time_out_seconds += seconds_in_a_day
-
-        time_difference = time_out_seconds - time_in_seconds
-
-        # Convert the difference to hours
-        work_duration_in_hours = time_difference / 3600
-
-        return round(work_duration_in_hours, 2)
 
     def getTypeOfDate(self, trans_date):
         try:
